Feasible in edge cases, program is extremely bare-boned now

+++ Combinatorisch/src/case2/Depot.java
@@ -1,5 +1,6 @@
 package case2;
 
+import java.io.BufferedWriter;
 import java.util.ArrayList;
 
 public class Depot {
@@ -9,17 +10,19 @@ public class Depot {
 	ArrayList<Tool> toolstack;
 	ArrayList<Vehicle> carpark;
 	int numVehicle = 0;
+//	BufferedWriter bf;
 
-	Depot(int id, int x, int y, int[][] tools, int maxCap, int maxDist) {
+	Depot(int id, int x, int y, long[][] tools, int maxCap, int maxDist) {
 		location = new Location(id, x, y);
 		this.maxCap = maxCap;
 		this.maxDist = maxDist;
 		toolstack = new ArrayList<>();
 		carpark = new ArrayList<>();
+//		this.bf = bf;
 		init(tools);
 	}
 
-	void init(int[][] tools) {
+	void init(long[][] tools) {
 		for (int i = 0; i < tools.length; i++) {
 			for (int j = 0; j < tools[i][2]; j++) {
 				toolstack.add(new Tool(tools[i][0], tools));
